fix: add namespace import support for JSON files
The namespace import feature was implemented for .mld files but was missing
for JSON imports. This adds the missing `importNamespace` case to the JSON
import handler, allowing syntax like:

    @import { * as config } from [config.json]

The imported JSON object is available as a namespace variable that can be
accessed with dot notation in templates:

    {{config.name}}, {{config.nested.value}}

Also updated test mappings to include the JSON namespace import test cases.